requirements.txt
aiogram==2.23.1
mysql-connector-python==8.1.0
mysql==0.0.3
mysqlclient==2.2.0
urllib3>=2.0.6
Pillow>=10.0.1
requests==2.31.0
python-dotenv~=1.0.0
beautifulsoup4~=4.12.2
reportlab~=4.0.6
python-docx~=1.0.0
# youtube-search-python==1.6.6
# pdf2docx~=0.5.6
# selenium~=4.13.0
